slm: fix fp16 overflow that silently set val_loss to inf
evaluate() builds logits inside amp_ctx (float16 on CUDA) but calls
F.cross_entropy outside it, so the per-token losses come back fp16 and
token_losses[valid].sum() accumulates in fp16 too. That sum is tokens x mean
nats; for the 12,455-token val set it passes fp16's 65504 max at a mean of
~5.26 nats, and val_loss — which drives checkpoint selection — becomes inf with
no warning. Cast the logits to float32 for that call.

Caught by an AdamW/Muon control pair that both reported val_ppl = exp(20), the
clamp for a non-finite loss. Published numbers survive (slm_A at 3.74 nats is
the closest, at 93% of the ceiling), but every run's first epochs were being
discarded, as would be any config worse than ppl 193.

Regression test: slm_training/tests/test_eval_fp16_overflow.py.
Same smoke pair shows Muon a full nat ahead of AdamW at step 40 (5.056 vs
6.099) at matched grad norms; full pretrain submitted.

+++ slm_training/src/train/run_small_lm.py
@@ -470,8 +470,13 @@ def evaluate(
                 out  = model(x, cond, y, cond_mask=mask)
             else:
                 out  = model(x, y)
+        # .float() is load-bearing: this call sits OUTSIDE amp_ctx, so under
+        # float16 AMP the logits arrive as fp16 and the per-token losses come back
+        # as fp16 too. Summing ~12k of them at ~5 nats overflows fp16's 65504 max,
+        # so val_loss silently becomes inf for any run above ~5.3 nats — which is
+        # every run early in training, and every larger validation set.
         token_losses = F.cross_entropy(
-            out.logits.reshape(-1, out.logits.size(-1)),
+            out.logits.reshape(-1, out.logits.size(-1)).float(),
             y.reshape(-1),
             ignore_index=-100,
             reduction="none",
